clean and improve code

+++ script/pre_immersive_distorted.py
@@ -28,13 +28,17 @@ import numpy as np
 import shutil
 import pickle
 import argparse
-
+import cv2
+import numpy as np
+import os 
+import json 
 
 import natsort 
 import sys
 import struct
 import pickle
 from scipy.spatial.transform import Rotation
+import sys 
 sys.path.append(".")
 from thirdparty.gaussian_splatting.utils.my_utils import posetow2c_matrcs, rotmat2qvec, qvec2rotmat
 from thirdparty.gaussian_splatting.utils.graphics_utils import focal2fov, fov2focal
@@ -43,8 +47,7 @@ from thirdparty.gaussian_splatting.helper3dg import getcolmapsingleimdistort
 from script.pre_n3d import extractframes
 SCALEDICT = {}
 
-# SCALEDICT["01_Welder_S11"] = 0.35
-# ["04_Truck", "09_Alexa", "10_Alexa", "11_Alexa", "12_Cave"]
+
 Immersiveseven = ["01_Welder",  "02_Flames", "04_Truck", "09_Alexa", "10_Alexa", "11_Alexa", "12_Cave"]
 immmersivescaledict = {}
 immmersivescaledict["01_Welder"] = 0.36
@@ -58,12 +61,6 @@ immmersivescaledict["12_Cave"] = 0.36
 for scene in Immersiveseven:
     immmersivescaledict[scene + "_dist"] =immmersivescaledict[scene] 
     SCALEDICT[scene + "_dist"] = immmersivescaledict[scene]  #immmersivescaledict[scene]  # to be checked with large scale
-
-
-for scene in Immersiveseven:
-    immmersivescaledict[scene + "_S14"] =immmersivescaledict[scene]  # inner datset name
-    SCALEDICT[scene + "_S14"] = immmersivescaledict[scene]  #immmersivescaledict[scene]  # to be checked with large scale
-
 
 
 def convertmodel2dbfiles(path, offset=0, scale=1.0, removeverythingexceptinput=False):
@@ -104,7 +101,7 @@ def convertmodel2dbfiles(path, offset=0, scale=1.0, removeverythingexceptinput=F
 
 
     import json 
-    with open(os.path.join(video + "models.json"), "r") as f:
+    with open(os.path.join(path, "models.json"), "r") as f:
         meta = json.load(f)
 
     for idx , camera in enumerate(meta):
@@ -268,7 +265,7 @@ def imageundistort(video, offsetlist=[0],focalscale=1.0, fixfocal=None):
     import numpy as np
     import os 
     import json 
-    with open(os.path.join(video + "models.json"), "r") as f:
+    with open(os.path.join(video, "models.json"), "r") as f:
                 meta = json.load(f)
 
     for idx , camera in enumerate(meta):
@@ -286,40 +283,46 @@ def imageundistort(video, offsetlist=[0],focalscale=1.0, fixfocal=None):
             videofolder = os.path.join(video, folder)
             imagepath = os.path.join(videofolder, str(offset) + ".png")
             imagesavepath = os.path.join(video, "colmap_" + str(offset), "input", folder + ".png")
-            
-            inputimagefolder = os.path.join(video, "colmap_" + str(offset), "input")
-            if not os.path.exists(inputimagefolder):
-                os.makedirs(inputimagefolder)
-            assert os.path.exists(imagepath)
-            image = cv2.imread(imagepath).astype(np.float32) #/ 255.0
-            h, w = image.shape[:2]
+            if os.path.exists(imagesavepath):
+                pass
+            else:
+                inputimagefolder = os.path.join(video, "colmap_" + str(offset), "input")
+                if not os.path.exists(inputimagefolder):
+                    os.makedirs(inputimagefolder)
+                assert os.path.exists(imagepath)
+                image = cv2.imread(imagepath).astype(np.float32) #/ 255.0
+                h, w = image.shape[:2]
 
 
-            image_size = (w, h)
-            knew = np.zeros((3, 3), dtype=np.float32)
+                image_size = (w, h)
+                knew = np.zeros((3, 3), dtype=np.float32)
 
- 
-            knew[0,0] = focalscale * intrinsics[0,0]
-            knew[1,1] = focalscale * intrinsics[1,1]
-            knew[0,2] =  view['principal_point'][0] # cx fixed half of the width
-            knew[1,2] =  view['principal_point'][1] #
-            knew[2,2] =  1.0
+    
+                knew[0,0] = focalscale * intrinsics[0,0]
+                knew[1,1] = focalscale * intrinsics[1,1]
+                knew[0,2] =  view['principal_point'][0] # cx fixed half of the width
+                knew[1,2] =  view['principal_point'][1] #
+                knew[2,2] =  1.0
 
 
-            
-            
-            map1, map2 = cv2.fisheye.initUndistortRectifyMap(intrinsics, dis_cef, R=None, P=knew, size=(w, h), m1type=cv2.CV_32FC1)
+           
+                map1, map2 = cv2.fisheye.initUndistortRectifyMap(intrinsics, dis_cef, R=None, P=knew, size=(w, h), m1type=cv2.CV_32FC1)
 
-            undistorted_image = cv2.remap(image, map1, map2, interpolation=cv2.INTER_CUBIC, borderMode=cv2.BORDER_CONSTANT)
-            undistorted_image = undistorted_image.clip(0,255.0).astype(np.uint8)
-
-            cv2.imwrite(imagesavepath, undistorted_image)
+                undistorted_image = cv2.remap(image, map1, map2, interpolation=cv2.INTER_CUBIC, borderMode=cv2.BORDER_CONSTANT)
+                undistorted_image = undistorted_image.clip(0,255.0).astype(np.uint8)
+           
+                cv2.imwrite(imagesavepath, undistorted_image)
 
             if offset == 0:
-                #
-                distortingflow = getdistortedflow(image, intrinsics, dis_cef, "linear", crop_output=False, scale=1.0, knew=knew)
-                print("saved distortion mappers")
-                np.save(os.path.join(video, folder  + ".npy"), distortingflow)
+                distortionmapperpath = os.path.join(video, folder  + ".npy")
+                 
+                if os.path.exists(distortionmapperpath):
+                    print("already exists mapper")
+                    pass 
+                else:
+                    distortingflow = getdistortedflow(image, intrinsics, dis_cef, "linear", crop_output=False, scale=1.0, knew=knew)
+                    print("saved distortion mappers")
+                    np.save(os.path.join(video, folder  + ".npy"), distortingflow)
 
 
 
@@ -334,17 +337,16 @@ def softlinkdataset(originalpath, path, srcscene, scene):
 
     for videofolder in videofolderlist:
         newlink = os.path.join(path, videofolder.split("/")[-2])
-        if os.path.exists(newlink):
+        if not os.path.exists(newlink):
+            cmd = " ln -s " + videofolder + " " + newlink
+            os.system(cmd)
+            print(cmd)
+        else:
             print("already exists do not make softlink again")
-            quit()
-        assert not os.path.exists(newlink)
-        cmd = " ln -s " + videofolder + " " + newlink
-        os.system(cmd)
-        print(cmd)
 
-    originalmodel = originalpath + "models.json"
-    newmodel = path + "models.json"
-    shutil.copy(originalmodel, newmodel)
+    originalmodel = os.path.join(originalpath, "models.json")
+    newmodelpath = os.path.join(path, "models.json")
+    shutil.copy(originalmodel, newmodelpath)
 if __name__ == "__main__" :
     parser = argparse.ArgumentParser()
  
@@ -388,8 +390,8 @@ if __name__ == "__main__" :
     scene = srcscene + postfix
     originalpath = videopath #" 
     originalvideo = originalpath# 43 1
-    path = videopath[:-1] + postfix
-    video = originalpath  # 43 1 
+    dstpath = videopath[:-1] + postfix # the path to save the dataset.
+
     scale = immmersivescaledict[scene]
 
 
@@ -397,23 +399,18 @@ if __name__ == "__main__" :
     for v in tqdm.tqdm(videoslist):
         extractframes(v)
 
-    try:
-        softlinkdataset(originalpath, path, srcscene, scene)
-    except:
-        print("softlink failed")
-        quit()
-    try:
-        imageundistort(video, offsetlist=[i for i in range(startframe,endframe)],focalscale=scale, fixfocal=None)
-    except:
-        print("undistort failed")
-        quit()
+    softlinkdataset(originalpath, dstpath, srcscene, scene)
+  
+    
+    imageundistort(dstpath, offsetlist=[i for i in range(startframe,endframe)],focalscale=scale, fixfocal=None)
+
 
     try:
         for offset in tqdm.tqdm(range(startframe, endframe)):
-            convertmodel2dbfiles(video, offset=offset, scale=scale, removeverythingexceptinput=False)
+            convertmodel2dbfiles(dstpath, offset=offset, scale=scale, removeverythingexceptinput=False)
     except:
-        convertmodel2dbfiles(video, offset=offset, scale=scale, removeverythingexceptinput=True)
+        convertmodel2dbfiles(dstpath, offset=offset, scale=scale, removeverythingexceptinput=True)
         print("create colmap input failed, better clean the data and try again")
         quit()
     for offset in range(startframe, endframe):
-        getcolmapsingleimdistort(video, offset=offset)
+        getcolmapsingleimdistort(dstpath, offset=offset)
